dsp_filters: add bitcrusher, overdrive, earlyreverb and wsolapitchtempo; native int16 in every filter
The DSP filter set gains a bitcrusher, an overdrive, an early-reflection
reverb and a WSOLA pitch shifter with their presets, and every filter
processes float and int16 natively, so an all-float or all-int16 chain
never converts samples. The four are registered in the builtin table in
dsp_filter.c, in griffin and in the static audio filter objects of
Makefile.common alongside the existing eleven.

Every filter builds in the strict C89 lane whether or not the C99 math
names are declared, which compile-matrix.sh checks for each filter as
consoles build it; eq's float output grows with the input instead of
overrunning a fixed buffer; and retro_dsp_filter_free() releases the plug
table in builtin-only builds as well.

+++ libretro-common/include/filters.h
@@ -97,7 +97,7 @@ static INLINE double besseli0(double x)
 
 static INLINE double kaiser_window_function(double index, double beta)
 {
-   return besseli0(beta * sqrtf(1 - index * index));
+   return besseli0(beta * sqrt(1 - index * index));
 }
 
 #endif
